api/garmin/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
garminconnect>=0.2.20
PyJWT>=2.8.0
pydantic>=2.0.0
